brave
adding support for the new Brave browser
Only thing is that the cache also is in the same folder.
doesn't matter to me, but maybe it will pain some but. you deside;)

first time doing this. don't know if this is the right way to propose changes. Please let me know.