Fix `surf` support (rm wrong '/')
This trailing slash prevents `psd` to form a proper dir-name for backup/overlay, so `psd` does not work properly with `surf`, issuing the message:
`stat: cannot statx '~/.surf/-backup': No such file or directory`.
`surf` keeps its config in "~/.surf". Before this fix, the resulting backup dir is "~/.surf/-backup", not "~/.surf-backup" as it should be.
This means that `psd` tries to create a backup inside the dir which it supposed to make backup of. :)
After removing this unfortunate '/', `psd` works with `surf` correctly, making proper dirs: '~/.surf-backup' and '~/.surf-back-ovfs'.
| sync target:  /home/elf/.surf
| tmpfs dir:    /run/user/1000/elf-surf

Note, that none of other browser's configs (in "/usr/share/psd/browsers/*") contains such trailing '/'.

Thanks for you work on this Daemon! : )